Use lower-case final "t" in "SComCat"
There had been inconsistent signals from upstream as to whether it's
"SComCaT" or "SComCat", but the preponderance of evidence does seem to
point at the latter (and there's no logical reason for the former that
I can see).  I already adjusted the repository's name in the GitHub
settings for our clone, excuse me I mean "fork", matching this change.